third_party/blink/web_tests/virtual/select-parser-relaxation/external/wpt/html/semantics/forms/the-select-element/customizable-select/selectedcontentelement-attr.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] The selectedcontentelement attribute should set up an association between a select and a selectedcontent.
  assert_equals: Setting the selectedcontentelement attribute via setAttribute should synchronously assign the contents of <selectedcontent>. expected "one<span>span</span>" but got ""
Harness: the test ran to completion.
